Change session names to reflect actual ones
Session names were originally based on rosbag names, but capture
session names are based on the start time found in the rosbag, which, if
the timing works out perfectly, is the next minute (the rosbag creating
script sets the name of the rosbag at 15:59:59 for example, and the
recording starts at 16:00:00), which makes the names different.
Corrected rosbag names to reflect actual start time of recordings.